Fractions without file read
Asks user for 2 fractions and automatically adds and multiplies them. In a while loop so the user can put in two more fractions if needed. 

Function for read_file is there but only reads in the first piece of data. Also do not know how to pass the two fractions from normalize() (where the fractions are put into lowest terms) to the bool function to see if they are equivalent.